ml-service/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
joblib
python-multipart
pdfminer.six
python-docx
